less enforcement & better structure

+++ src/main/java/net/unknownuser/eventreader/codes/SyncEvents.java
@@ -6,8 +6,8 @@ public abstract class SyncEvents {
 	 * e.g.: a mouse can send REL_X and REL_Y signals followed by REPORT. The REL_X and REL_Y are
 	 * separate events, but belong to the same movement.
 	 */
-	public static final byte REPORT = 0x0;
-	public static final byte CONFIG = 0x1;
-	public static final byte MULTITOUCH_REPORT = 0x2;
-	public static final byte DROPPED = 0x3;
+	public static final short REPORT = 0x0;
+	public static final short CONFIG = 0x1;
+	public static final short MULTITOUCH_REPORT = 0x2;
+	public static final short DROPPED = 0x3;
 }
